pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anselm-experiment"
version = "0.1.0"
description = "Empirical harness for ANSELM Article 4 §10 — accumulation vs. averaging across agent architectures."
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.50",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "jsonschema>=4.22",
    "bert-score>=0.3.13",
    "rich>=13.7",
    "python-dotenv>=1.0",
    "tiktoken>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "ruff>=0.5",
    "jupyter>=1.0",
    "matplotlib>=3.9",
    "pandas>=2.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
